refactor: рефакторинг страницы Profile по замечаниям
